--- src/background/auth.ts
@@ -1,32 +1,0 @@
-import { Storage } from "@plasmohq/storage"
-import { logModule } from "./service/log"
-const storage = new Storage()
-
-/**
- * @param url 요청 URL
- * @param method 요청 메서드
- * @param body 요청 본문
- * @returns 응답 데이터 (any)
- */
-export async function Request(
-  url: string,
-  method: "GET" | "POST" | "PUT" | "DELETE",
-  body: any = {}
-): Promise<any | null> {
-  logModule.devLog("Request")
-  try {
-    const ret = await fetch(url, {
-      method,
-      headers: {
-        accept: "*/*",
-        "Content-Type": "application/json"
-      },
-      body: method === "GET" ? undefined : JSON.stringify(body)
-    })
-
-    if (ret.ok) return await ret.json()
-    else return logModule.log("error", ret.statusText)
-  } catch (e) {
-    return logModule.log("error", e)
-  }
-}
