+++ src/background/messages/page.ts
@@ -1,5 +1,5 @@
 import type { PlasmoMessaging } from "@plasmohq/messaging"
-import { modifyTab, newTab } from "~background/service/page"
+import { modifyTab, newTab } from "~background/page"
 
 const handler: PlasmoMessaging.MessageHandler = async (req, res) => {
   const { msg } = req.body
